requirements.txt
asgiref==3.7.2
cachetools==4.2.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.2
django-allauth==0.61.1
djangorestframework==3.14.0
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.48.2
idna==3.6
oauthlib==3.2.2
protobuf==3.20.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
